# [ROX-12211] Optimize image graphQL resolver