--- a/central/graphql/resolvers/image_scan.go
+++ b/central/graphql/resolvers/image_scan.go
@@ -80,8 +80,12 @@ func getImageComponentResolvers(ctx context.Context, root *Resolver, imageScan *
 			return resolvers[i].data.GetId() < resolvers[j].data.GetId()
 		})
 	}
+	resolverI := make([]ImageComponentResolver, 0, len(resolvers))
+	for _, resolver := range resolvers {
+		resolverI = append(resolverI, resolver)
+	}
 	ret, err := paginationWrapper{
 		pv: query.GetPagination(),
-	}.paginate(resolvers, nil)
+	}.paginate(resolverI, nil)
 	return ret.([]ImageComponentResolver), err
 }