## Description
Optimize image with scan and image without scan graphQL paths. In the past we always loaded the full image even when scan details were not needed. This PR optimizes boths paths for the new graphQL endpoints.

Note: It would be easier to review individual commits.


## Checklist
- [ ] Investigated and inspected CI test results
- [X] Unit test and regression tests added
- ~Evaluated and added CHANGELOG entry if required~
- ~Determined and documented upgrade steps~
- ~Documented user facing changes~

If any of these don't apply, please comment below.

## Testing Performed
```
BenchmarkImageResolver
    env_isolator.go:41: EnvIsolator: Setting ROX_POSTGRES_DATASTORE to true
BenchmarkImageResolver/GetImageComponentsInImageScanResolver
BenchmarkImageResolver/GetImageComponentsInImageScanResolver-8         	      19	  58440569 ns/op
BenchmarkImageResolver/GetImageComponentsWithoutImageScanResolver
BenchmarkImageResolver/GetImageComponentsWithoutImageScanResolver-8    	       3	 441044677 ns/op
BenchmarkImageResolver/GetImageComponentsDerivedFieldsWithImageScanResolver
BenchmarkImageResolver/GetImageComponentsDerivedFieldsWithImageScanResolver-8         	      12	 132206259 ns/op
BenchmarkImageResolver/GetImageComponentsDerivedWithoutImageScanResolver
BenchmarkImageResolver/GetImageComponentsDerivedWithoutImageScanResolver-8            	       2	 619884703 ns/op
BenchmarkImageResolver/GetImageOnly
BenchmarkImageResolver/GetImageOnly-8                                                 	    2704	    475802 ns/op
BenchmarkImageResolver/GetImageWithCounts
BenchmarkImageResolver/GetImageWithCounts-8                                           	      16	  71430526 ns/op
PASS
```

<img width="1539" alt="image" src="https://user-images.githubusercontent.com/9895898/184973651-361dddcd-a6ae-4646-91df-1c5b8fbf1b09.png">

<img width="1539" alt="image" src="https://user-images.githubusercontent.com/9895898/184973703-4c724829-0a63-4e3b-90b4-5588652964e7.png">